adjust dependencies for forward compatibility
When installing this component in React 18, using TypeScript, warnings appear
in your editor since this package depends on `@types/react` directly. This is
only necessary for the build environment, and shouldn't be included in a
package's dependencies as it can confuse TypeScript at the application level.

This adjustment to the dependencies still lints and builds just fine, but the
package itself will no longer depend on anything at runtime, allowing usage of
the component in any future version of React.